requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
SQLAlchemy>=2.0.36
pymysql==1.1.0
python-dotenv==1.0.0
Werkzeug==3.0.1
Flask-Cors==3.0.10

